+++ plugins/owner/backup.js
@@ -1,4 +1,4 @@
-import JSZip from 'jszip'
+import { ZipFile as JSZip } from '../../lib/converter.js'
 import { join } from 'path'
 import { statSync, readFileSync, readdirSync } from 'fs'